pyproject.toml
[project]
name = "cwa"
version = "0.1.0"
description = ""
authors = [
    {name = "jerry.liu",email = "jerry.liu@uancare.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "llama-cpp-python (>=0.3.12,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

